Rename ADU *device_information, send update result details
The ADU documentation refers to the device basic information as
"device properties", not "device information", so this change renames
the structures and variables to match that naming.
Also the ADU sample gets an improvement to send a description for the
update result details, as well as an user-defined extended result code.